# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  user.CreateRequest:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
      profile_photo:
        type: string
      role:
        type: integer
    required:
    - email
    - name
    - password
    type: object
  volunteer.CreateRequest:
    properties:
      attribute_volunteer_id:
        minimum: 1
        type: integer
      birth_date:
        type: string
      birth_place:
        maxLength: 100
        type: string
      blood_type:
        enum:
        - A
        - B
        - AB
        - O
        type: string
      email:
        maxLength: 100
        type: string
      field:
        maxLength: 100
        type: string
      free_times:
        items:
          $ref: '#/definitions/volunteer.FreeTimeRequest'
        minItems: 1
        type: array
      home_phone:
        maxLength: 30
        type: string
      indonesian_name:
        maxLength: 100
        type: string
      interested_activity_ids:
        items:
          type: integer
        minItems: 1
        type: array
      languages:
        maxLength: 255
        type: string
      last_education:
        enum:
        - SD
        - SMP
        - SMA
        - D1
        - D2
        - D3
        - D4
        - S1
        - S2
        - S3
        type: string
      level_volunteer_id:
        minimum: 1
        type: integer
      mandarin_name:
        maxLength: 100
        type: string
      marital_status:
        enum:
        - single
        - married
        - divorced
        - widowed
        type: string
      master_area_id:
        minimum: 1
        type: integer
      office_phone:
        maxLength: 30
        type: string
      phone:
        maxLength: 30
        type: string
      postal_code:
        maxLength: 20
        type: string
      private_vehicle:
        type: boolean
      profession:
        maxLength: 100
        type: string
      regular_donor:
        type: boolean
      religion_id:
        minimum: 1
        type: integer
      residential_address:
        maxLength: 255
        type: string
      rhesus:
        enum:
        - positive
        - negative
        - unknown
        type: string
      vis_id:
        maxLength: 50
        type: string
    required:
    - attribute_volunteer_id
    - birth_date
    - birth_place
    - blood_type
    - email
    - field
    - free_times
    - indonesian_name
    - interested_activity_ids
    - languages
    - last_education
    - level_volunteer_id
    - marital_status
    - master_area_id
    - phone
    - postal_code
    - profession
    - religion_id
    - residential_address
    - rhesus
    - vis_id
    type: object
  volunteer.FreeTimeRequest:
    properties:
      day_of_week:
        enum:
        - monday
        - tuesday
        - wednesday
        - thursday
        - friday
        - saturday
        - sunday
        type: string
      period:
        enum:
        - morning
        - afternoon
        - evening
        type: string
    required:
    - day_of_week
    - period
    type: object
  volunteer.UpdateRequest:
    properties:
      attribute_volunteer_id:
        minimum: 1
        type: integer
      birth_date:
        type: string
      birth_place:
        maxLength: 100
        type: string
      blood_type:
        enum:
        - A
        - B
        - AB
        - O
        type: string
      email:
        maxLength: 100
        type: string
      field:
        maxLength: 100
        type: string
      free_times:
        items:
          $ref: '#/definitions/volunteer.FreeTimeRequest'
        minItems: 1
        type: array
      home_phone:
        maxLength: 30
        type: string
      indonesian_name:
        maxLength: 100
        type: string
      interested_activity_ids:
        items:
          type: integer
        minItems: 1
        type: array
      languages:
        maxLength: 255
        type: string
      last_education:
        enum:
        - SD
        - SMP
        - SMA
        - D1
        - D2
        - D3
        - D4
        - S1
        - S2
        - S3
        type: string
      level_volunteer_id:
        minimum: 1
        type: integer
      mandarin_name:
        maxLength: 100
        type: string
      marital_status:
        enum:
        - single
        - married
        - divorced
        - widowed
        type: string
      master_area_id:
        minimum: 1
        type: integer
      office_phone:
        maxLength: 30
        type: string
      phone:
        maxLength: 30
        type: string
      postal_code:
        maxLength: 20
        type: string
      private_vehicle:
        type: boolean
      profession:
        maxLength: 100
        type: string
      regular_donor:
        type: boolean
      religion_id:
        minimum: 1
        type: integer
      residential_address:
        maxLength: 255
        type: string
      rhesus:
        enum:
        - positive
        - negative
        - unknown
        type: string
      vis_id:
        maxLength: 50
        type: string
    required:
    - attribute_volunteer_id
    - birth_date
    - birth_place
    - blood_type
    - email
    - field
    - free_times
    - indonesian_name
    - interested_activity_ids
    - languages
    - last_education
    - level_volunteer_id
    - marital_status
    - master_area_id
    - phone
    - postal_code
    - profession
    - religion_id
    - residential_address
    - rhesus
    - vis_id
    type: object
info:
  contact: {}
  description: API documentation
  title: Social BE API
  version: "1.0"
paths:
  /admin/users:
    post:
      consumes:
      - application/json
      description: Create new user account by superadmin
      parameters:
      - description: Create user request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - users
...
  /users:
    get:
      description: Get list of users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all users
      tags:
      - users
  /users/{id}:
    get:
      description: Get detail user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get user by ID
      tags:
      - users
  /users/me:
    get:
      description: Get authenticated user profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - users
  /users/upload-photo:
    post:
      consumes:
      - multipart/form-data
      description: Upload profile photo to NAS and return URL
      parameters:
      - description: Profile photo
        in: formData
        name: profile_photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Upload profile photo
      tags:
      - users
  /volunteers:
    get:
      description: Get paginated list of volunteer
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all volunteer
      tags:
      - volunteer
    post:
      consumes:
      - application/json
      description: Create new volunteer
      parameters:
      - description: Create volunteer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/volunteer.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create volunteer
      tags:
      - volunteer
  /volunteers/{id}:
    delete:
      description: Soft delete volunteer
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete volunteer
      tags:
      - volunteer
    get:
      description: Get detail volunteer
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get volunteer by ID
      tags:
      - volunteer
    put:
      consumes:
      - application/json
      description: Update existing volunteer
      parameters:
      - description: Volunteer ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update volunteer
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/volunteer.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update volunteer
      tags:
      - volunteer
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"